Add delay between stderr and stdout output
Output to `stderr` and output to `stdout` is frequently out-of-order
(i.e., not synchronized) on adb shell non-interactive mode.

This problem is never happen on adb shell interactive mode,
so it seems like an `adb` issue.
Similar issue: https://stackoverflow.com/questions/51264320/cout-and-cerr-not-synchronized

To work around the issue, this patch adds 5ms delay between
`stderr` output and `stdout` output.

Tested 100000 times on Android 12 (SP1A.210812.015) of Pixel 3 XL.